pyproject.toml
[tool.poetry]
name = "in-the-game"
version = "0.1.0"
description = ""
authors = ["Kamil Podbielski <freeniac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
databases = "^0.4.1"
asyncpg = "^0.21.0"
psycopg2-binary = "^2.8.6"
PyJWT = "^2.0.1"
pydantic = {extras = ["email"], version = "^1.7.3"}

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
flake8 = "^3.8.4"
black = "^20.8b1"
bandit = "^1.7.0"
mypy = "^0.800"
requests = "^2.25.1"
alembic = "^1.5.2"
httpie = "^2.3.0"
isort = "^5.7.0"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128
target-version = ['py39']

[tool.isort]
profile = "black"
src_paths = "in_the_game,migrations,tests"
